feat(TDI-41801):extract common runtime codes from Mongodb.
