+++ mongo/mongo-common/pom.xml
@@ -5,7 +5,7 @@
     <parent>
         <groupId>org.talend.components</groupId>
         <artifactId>mongo</artifactId>
-        <version>1.26.0-SNAPSHOT</version>
+        <version>1.29.0-SNAPSHOT</version>
     </parent>
 
     <artifactId>mongo-common</artifactId>
@@ -17,6 +17,12 @@
             <groupId>org.mongodb</groupId>
             <artifactId>mongo-java-driver</artifactId>
             <version>${mongodb.client.version}</version>
+        </dependency>
+
+        <dependency>
+            <groupId>org.talend.components</groupId>
+            <artifactId>stream-json</artifactId>
+            <version>${project.version}</version>
         </dependency>
     </dependencies>
 </project>
