+++ mongo/mongo-common/src/main/java/org/talend/components/mongo/PathMapping.java
@@ -10,7 +10,7 @@
  * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations under the License.
  */
-package org.talend.components.mongodb;
+package org.talend.components.mongo;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
