+++ mongo/mongo-common/src/main/java/org/talend/components/mongo/service/I18nMessage.java
@@ -10,16 +10,14 @@
  * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations under the License.
  */
-package org.talend.components.mongodb.source;
+package org.talend.components.mongo.service;
 
-import org.talend.components.mongodb.dataset.BaseDataSet;
+import org.talend.sdk.component.api.internationalization.Internationalized;
 
-import java.io.Serializable;
+@Internationalized
+public interface I18nMessage {
 
-public interface BaseSourceConfiguration extends Serializable {
+    String example(String p1, String p2);
 
-    BaseDataSet getDataset();
-
-    Long getSampleLimit();
-
+    String missingConnection();
 }
