+++ mongo/mongo-common/src/main/resources/org/talend/components/mongo/Messages.properties
@@ -43,7 +43,11 @@ WriteConcern.JOURNALED._displayName=Journaled
 WriteConcern.REPLICA_ACKNOWLEDGED._displayName=Replica acknowledged
 WriteConcern.UNACKNOWLEDGED._displayName=Unacknowledged
 
-DataAction.DELETE._displayName=Delete
-DataAction.INSERT._displayName=Insert
-DataAction.SET._displayName=Set
-DataAction.UPSERT_WITH_SET._displayName=Upsert with set
+PathMapping.column._displayName=Field
+PathMapping.column._placeholder=
+
+PathMapping.originElement._displayName=Origin element
+PathMapping.originElement._placeholder=
+
+PathMapping.parentNodePath._displayName=Parent node path
+PathMapping.parentNodePath._placeholder=
