--- mongo/mongo-common/src/main/resources/org/talend/components/mongo/dataset/Messages.properties
@@ -1,5 +1,0 @@
-MongoCommonDataSet.collection._displayName=Collection
-MongoCommonDataSet.collection._placeholder=
-
-MongoCommonDataSet.mode._displayName=Mode
-MongoCommonDataSet.mode._placeholder=
