--- mongo/mongo-common/src/main/resources/org/talend/components/mongo/datastore/Messages.properties
@@ -1,17 +1,0 @@
-MongoCommonDataStore.addressType._displayName=Address type
-MongoCommonDataStore.addressType._placeholder=
-
-MongoCommonDataStore.address._displayName=Address
-MongoCommonDataStore.address._placeholder=
-
-MongoCommonDataStore.replicaSetAddress._displayName=ReplicaSet address
-MongoCommonDataStore.replicaSetAddress._placeholder=
-
-MongoCommonDataStore.database._displayName=Database
-MongoCommonDataStore.database._placeholder=
-
-MongoCommonDataStore.auth._displayName=Authentication
-MongoCommonDataStore.auth._placeholder=
-
-MongoCommonDataStore.connectionParameter._displayName=Connection parameter
-MongoCommonDataStore.connectionParameter._placeholder=
