+++ mongo/mongo-common/src/main/resources/org/talend/components/mongo/service/Messages.properties
@@ -1,0 +1,2 @@
+org.talend.components.mongo.service.I18nMessage.example=Example : {0},{1}
+org.talend.components.mongo.service.I18nMessage.missingConnection=Cannot find connection object for close
