+++ mongo/mongodb/pom.xml
@@ -10,7 +10,7 @@
 
   <artifactId>mongodb</artifactId>
 
-  <name>Connectors SE :: MongoDB</name>
+  <name>Connectors SE :: Mongo :: MongoDB</name>
 
   <dependencies>
     <!--
@@ -59,7 +59,7 @@
     <dependency>
         <groupId>org.talend.components</groupId>
         <artifactId>mongo-common</artifactId>
-        <version>1.26.0-SNAPSHOT</version>
+        <version>${project.version}</version>
         <scope>compile</scope>
     </dependency>
