+++ mongo/mongodb/src/main/java/org/talend/components/mongodb/DataAction.java
@@ -10,12 +10,14 @@
  * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations under the License.
  */
-package org.talend.components.mongodb.service;
+package org.talend.components.mongodb;
 
-import org.talend.sdk.component.api.internationalization.Internationalized;
-
-@Internationalized
-public interface I18nMessage {
-
-    String example(String p1, String p2);
+// hide update, upsert and delete as they don't support parallize
+public enum DataAction {
+    INSERT,
+    // UPDATE,
+    SET,
+    // UPSERT,
+    UPSERT_WITH_SET
+    // DELETE
 }
