+++ mongo/mongodb/src/main/java/org/talend/components/mongodb/datastore/MongoDBDataStore.java
@@ -12,20 +12,83 @@
  */
 package org.talend.components.mongodb.datastore;
 
+import lombok.Data;
+import org.talend.components.mongo.Address;
+import org.talend.components.mongo.AddressType;
+import org.talend.components.mongo.Auth;
+import org.talend.components.mongo.ConnectionParameter;
 import org.talend.components.mongo.datastore.MongoCommonDataStore;
 import org.talend.sdk.component.api.component.Icon;
 import org.talend.sdk.component.api.component.Version;
+import org.talend.sdk.component.api.configuration.Option;
 import org.talend.sdk.component.api.configuration.action.Checkable;
+import org.talend.sdk.component.api.configuration.condition.ActiveIf;
+import org.talend.sdk.component.api.configuration.constraint.Required;
 import org.talend.sdk.component.api.configuration.type.DataStore;
+import org.talend.sdk.component.api.configuration.ui.layout.GridLayout;
 import org.talend.sdk.component.api.meta.Documentation;
 
 import java.io.Serializable;
+import java.util.Collections;
+import java.util.List;
 
 @Version(1)
+@Data
 @Icon(value = Icon.IconType.CUSTOM, custom = "mongo_db-connector")
 @DataStore("MongoDBDataStore")
 @Checkable("healthCheck")
+@GridLayout(names = GridLayout.FormType.MAIN,
+        value = { @GridLayout.Row({ "addressType" }), @GridLayout.Row({ "address" }),
+                @GridLayout.Row({ "replicaSetAddress" }), @GridLayout.Row({ "database" }),
+                @GridLayout.Row({ "auth" }) })
+@GridLayout(names = GridLayout.FormType.ADVANCED, value = { @GridLayout.Row({ "connectionParameter" }) })
 @Documentation("MongoDB connection")
-public class MongoDBDataStore extends MongoCommonDataStore implements Serializable {
+public class MongoDBDataStore implements MongoCommonDataStore {
 
+    @Option
+    @Required
+    @Documentation("https://docs.mongodb.com/manual/reference/connection-string/")
+    private AddressType addressType = AddressType.STANDALONE;
+
+    @Option
+    @ActiveIf(target = "addressType", value = "STANDALONE")
+    @Documentation("https://docs.mongodb.com/manual/reference/connection-string/")
+    private Address address;
+
+    @Option
+    @ActiveIf(target = "addressType", value = "REPLICA_SET")
+    @Documentation("https://docs.mongodb.com/manual/reference/connection-string/")
+    private List<Address> replicaSetAddress = Collections.emptyList();
+
+    /*
+     * @Option
+     *
+     * @ActiveIf(target = "addressType", value = "SHARDED_CLUSTER")
+     *
+     * @Documentation("https://docs.mongodb.com/manual/reference/connection-string/")
+     * private List<Address> shardedClusterAddress = Collections.emptyList();
+     */
+
+    @Option
+    @Required
+    @Documentation("Database")
+    private String database;
+
+    // TODO have to support the function for locate a file for cert if need cert? or trust any cert, not good for me?
+    // maybe remove it
+    /*
+     * @Option
+     *
+     * @Documentation("Use SSL")
+     * private boolean useSSL;
+     */
+
+    @Option
+    @Documentation("auth page")
+    private Auth auth;
+
+    // advanced page
+    @Option
+    @Documentation("Connection parameter")
+    private List<ConnectionParameter> connectionParameter = Collections.emptyList();
 }
