+++ mongo/mongodb/src/main/resources/org/talend/components/mongodb/Messages.properties
@@ -14,14 +14,10 @@ SortBy.column._placeholder=
 SortBy.order._displayName=Order
 SortBy.order._placeholder=
 
-PathMapping.column._displayName=Field
-PathMapping.column._placeholder=
-
-PathMapping.originElement._displayName=Origin element
-PathMapping.originElement._placeholder=
-
-PathMapping.parentNodePath._displayName=Parent node path
-PathMapping.parentNodePath._placeholder=
-
 AggregationStage.stage._displayName=Stage
 AggregationStage.stage._placeholder=
+
+DataAction.DELETE._displayName=Delete
+DataAction.INSERT._displayName=Insert
+DataAction.SET._displayName=Set
+DataAction.UPSERT_WITH_SET._displayName=Upsert with set
