+++ mongo/mongodb/src/main/resources/org/talend/components/mongodb/dataset/Messages.properties
@@ -1,6 +1,18 @@
 MongoDBReadDataSet.datastore._displayName=MongoDB
 
+MongoDBReadDataSet.collection._displayName=Collection
+MongoDBReadDataSet.collection._placeholder=
+
+MongoDBReadDataSet.mode._displayName=Mode
+MongoDBReadDataSet.mode._placeholder=
+
 MongoDBReadDataSet.query._displayName=Query
 MongoDBReadDataSet.query._placeholder=
 
 MongoDBReadAndWriteDataSet.datastore._displayName=MongoDB
+
+MongoDBReadAndWriteDataSet.collection._displayName=Collection
+MongoDBReadAndWriteDataSet.collection._placeholder=
+
+MongoDBReadAndWriteDataSet.mode._displayName=Mode
+MongoDBReadAndWriteDataSet.mode._placeholder=
