+++ mongo/mongodb/src/main/resources/org/talend/components/mongodb/datastore/Messages.properties
@@ -1,0 +1,17 @@
+MongoDBDataStore.addressType._displayName=Address type
+MongoDBDataStore.addressType._placeholder=
+
+MongoDBDataStore.address._displayName=Address
+MongoDBDataStore.address._placeholder=
+
+MongoDBDataStore.replicaSetAddress._displayName=ReplicaSet address
+MongoDBDataStore.replicaSetAddress._placeholder=
+
+MongoDBDataStore.database._displayName=Database
+MongoDBDataStore.database._placeholder=
+
+MongoDBDataStore.auth._displayName=Authentication
+MongoDBDataStore.auth._placeholder=
+
+MongoDBDataStore.connectionParameter._displayName=Connection parameter
+MongoDBDataStore.connectionParameter._placeholder=
