+++ mongo/pom.xml
@@ -5,7 +5,7 @@
     <parent>
         <groupId>org.talend.components</groupId>
         <artifactId>connectors-se</artifactId>
-        <version>1.26.0-SNAPSHOT</version>
+        <version>1.29.0-SNAPSHOT</version>
     </parent>
 
     <artifactId>mongo</artifactId>
